+++ SpaceShoter/ExEn 1.0.2/BlankGame/Properties/Entities/Interactable/Enemies/Enemy.cs
@@ -15,7 +15,7 @@ namespace BlankGame
 				//Sprite image;
 				protected Random r;
 				protected int points=30;
-				public Enemy(Game g, Vector2 pos,Vector2 direct,int timer)
+				public Enemy(Game g, Vector2 pos,Vector2 direct,float timer)
 				:base(g,timer)
 				{
 					this.direct = direct;
